core-service/pyproject.toml
[project]
name = "task-tracker-todo-django"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniil",email = "dan.shumski@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=44.0.3,<45.0.0)",
    "drf-nested-routers (>=0.94.2,<0.95.0)",
    "boto3 (>=1.38.36,<2.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "kafka-python (>=2.2.12,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
mypy = "^1.15.0"
isort = "^6.0.1"
black = "^25.1.0"
pre-commit = "^4.2.0"

[tool.poetry.group.prod.dependencies]
gunicorn = "^23.0.0"

[tool.poetry]
package-mode = false
